Refactor and potential fix to #11
Requires to manually set hidden attribute on windows. On *nix, this is
inherit in files that have a preceeding '.'. Using `Set-ItemProperty`
and setting the Hidden attribute is a no-op on Linux, thus it might be
safe to use on Linux and Windows. Will check when I get to a windows
machine next